requirements.txt
requests
pytest
python-decouple